MultilingualAIHealthCompanion/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
httpx==0.25.2
google-generativeai==0.3.1
elevenlabs==0.2.27
numpy==1.24.3
PyPDF2==3.0.1